pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.17.2",
    "fastapi[standard]>=0.128.0",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "sqlalchemy>=2.0.45",
    "uvicorn[standard]>=0.40.0",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
    "ruff>=0.14.10",
]

[tool.ruff]
line-length = 88
extend-exclude = [
    "htmlcov",
    "alembic",
]

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]

[tool.mypy]
python_version = "3.14"
show_error_codes = true
pretty = true
warn_unused_ignores = true
no_implicit_optional = true
ignore_missing_imports = true
exclude = [
    "config\\.py",
    "app/config\\.py",
]

[tool.pytest.ini_options]
# pytest自体のオプション (例: 詳細表示)
addopts = "-v --cov --cov-report=term-missing --cov-report=html"
#addopts = "-v"
# テストファイルがあるディレクトリ
testpaths = [ "tests"]
[tool.coverage.run]
source = ["app"]
branch = true
[tool.coverage.report]
# skip_covered = true
[tool.coverage.html]
# directory = "htmlcov"
